Data Validation: URL Constraints & Data Annotation
For further details, read:   aboutDataValidation.txt

+++ api/Controllers/CommentController.cs
@@ -28,41 +28,54 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepo.GetAllAsync();
             var commentDto = comments.Select(c => c.ToCommentDto());
-            
+
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")] // URL Constraint ({id:int}) = Data Validation 
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) // ModelState.IsValid = Makes the data annotations work (Another Data Validation)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.GetByIdAsync(id);
 
-            if(comment == null){
+            if (comment == null)
+            {
                 return NotFound();
             }
 
             return Ok(comment.ToCommentDto()); // return one
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto) // 1
         {
-            if(!await _stockRepo.StockExists(stockId))  // 2
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            if (!await _stockRepo.StockExists(stockId))  // 2
             {
                 return BadRequest("Stock does not exist");
             }
 
             var commentModel = commentDto.ToCommentFromCreate(stockId); // 1
             await _commentRepo.CreateAsync(commentModel);  // 3 and 4
-            return CreatedAtAction(nameof(GetById) , new { id = commentModel.Id}, commentModel.ToCommentDto()); // 5
+            return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto()); // 5
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if (comment == null)
@@ -74,12 +87,15 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var commentModel = await _commentRepo.DeleteAsync(id);
 
-            if(commentModel == null)
+            if (commentModel == null)
             {
                 return NotFound("Comment does not exist");
             }
